+++ app/src/main/java/com/polar/nextcloudservices/SettingsActivity.java
@@ -2,91 +2,30 @@ package com.polar.nextcloudservices;
 
 import java.util.List;
 import java.util.Timer;
-import java.util.TimerTask;
 
-import android.content.ComponentName;
-import android.content.ServiceConnection;
-import android.net.Uri;
 import android.os.Bundle;
 import android.os.Handler;
-import android.os.IBinder;
 import android.util.Log;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.Context;
 import android.app.ActivityManager;
 import android.view.MenuItem;
-import android.widget.Toast;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.DialogFragment;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
-import androidx.preference.EditTextPreference;
-import androidx.preference.PreferenceFragmentCompat;
 import androidx.preference.PreferenceManager;
-import androidx.preference.Preference;
 
-import com.nextcloud.android.sso.AccountImporter;
-import com.nextcloud.android.sso.api.NextcloudAPI;
-import com.nextcloud.android.sso.exceptions.AccountImportCancelledException;
-import com.nextcloud.android.sso.exceptions.AndroidGetAccountsPermissionNotGranted;
-import com.nextcloud.android.sso.exceptions.NextcloudFilesAppNotInstalledException;
-import com.nextcloud.android.sso.model.SingleSignOnAccount;
-import com.nextcloud.android.sso.ui.UiExceptionManager;
+import com.polar.nextcloudservices.settings.NotificationServiceConnection;
+import com.polar.nextcloudservices.settings.PreferenceUpdateTimerTask;
+import com.polar.nextcloudservices.ui.settings.SettingsFragment;
 
-import nl.invissvenska.numberpickerpreference.NumberDialogPreference;
-import nl.invissvenska.numberpickerpreference.NumberPickerPreferenceDialogFragment;
-
-
-class NotificationServiceConnection implements ServiceConnection {
-    private final String TAG = "SettingsActivity.NotificationServiceConnection";
-    private final SettingsActivity.SettingsFragment settings;
-    private NotificationService.Binder mService;
-    public boolean isConnected = false;
-
-    public NotificationServiceConnection(SettingsActivity.SettingsFragment _settings) {
-        settings = _settings;
-    }
-
-    @Override
-    public void onServiceConnected(ComponentName name, IBinder service) {
-        if (service instanceof NotificationService.Binder) {
-            mService = (NotificationService.Binder) service;
-            settings.setStatus(((NotificationService.Binder) service).getServiceStatus());
-            isConnected = true;
-        } else {
-            Log.wtf(TAG, "Bad Binder type passed!");
-            throw new RuntimeException("Expected NotificationService.Binder");
-        }
-    }
-
-    public void updateStatus(){
-        if(!isConnected){
-            Log.w(TAG, "Service has already disconnected");
-            settings.setStatus("Disconnected: service is not running");
-        } else {
-            settings.setStatus(mService.getServiceStatus());
-        }
-    }
-
-    public void tellAccountChanged(){
-        Log.d(TAG, "Telling service that account has cahnged");
-        mService.onAccountChanged();
-    }
-
-    @Override
-    public void onServiceDisconnected(ComponentName name) {
-        Log.w(TAG, "Service has disconnected.");
-        isConnected = false;
-    }
-}
 
 public class SettingsActivity extends AppCompatActivity implements SharedPreferences.OnSharedPreferenceChangeListener {
     private final String TAG = "SettingsActivity";
-    private final Handler mHandler = new Handler();
+    public final Handler mHandler = new Handler();
     private Timer mTimer = null;
     private PreferenceUpdateTimerTask mTask = null;
     private NotificationServiceConnection mServiceConnection = null;
@@ -100,31 +39,6 @@ public class SettingsActivity extends AppCompatActivity implements SharedPrefere
             return true;
         }
         return super.onOptionsItemSelected(item);
-    }
-
-    class PreferenceUpdateTimerTask extends TimerTask {
-        private final SettingsFragment settings;
-        public PreferenceUpdateTimerTask(SettingsFragment _settings) {
-            settings = _settings;
-        }
-
-        @Override
-        public void run() {
-            // run on another thread
-            mHandler.post(() -> {
-                //Log.d(TAG, "Entered run in preference updater timer task");
-                if(!getBoolPreference("enable_polling", true)){
-                    stopNotificationService();
-                } else if(!isNotificationServiceRunning()) {
-                    startNotificationService();
-                }
-                if (isNotificationServiceRunning()) {
-                    updateNotificationServiceStatus(settings);
-                } else {
-                    (settings).setStatus("Disconnected: service is not running");
-                }
-            });
-        }
     }
 
     public void stopNotificationService() {
@@ -245,7 +159,7 @@ public class SettingsActivity extends AppCompatActivity implements SharedPrefere
                 }
             }
             Log.d(TAG, "Starting timer");
-            mTask = new PreferenceUpdateTimerTask((SettingsFragment) settings);
+            mTask = new PreferenceUpdateTimerTask(this, (SettingsFragment) settings);
             mTimer.scheduleAtFixedRate( mTask, 0, 5000);
         }
     }
@@ -253,202 +167,6 @@ public class SettingsActivity extends AppCompatActivity implements SharedPrefere
     private void setupSharedPreferences() {
         SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
         sharedPreferences.registerOnSharedPreferenceChangeListener(this);
-    }
-
-
-
-    public static class SettingsFragment extends PreferenceFragmentCompat {
-        private final String TAG = "SettingsActivity.SettingsFragment";
-
-        private boolean getBoolPreference(String key, boolean fallback) {
-            SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(getContext());
-            return sharedPreferences.getBoolean(key, fallback);
-        }
-
-        private void notifyService(){
-            SettingsActivity activity = (SettingsActivity) getActivity();
-            if(activity == null){
-                Log.wtf(TAG, "Activity can not be null!");
-                throw new NullPointerException();
-            }
-            activity.mServiceConnection.tellAccountChanged();
-        }
-
-        private void enableSSO(@NonNull SingleSignOnAccount account){
-            SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(getContext());
-            SharedPreferences.Editor editor = sharedPreferences.edit();
-            editor.putBoolean("sso_enabled", true);
-            editor.putString("sso_name", account.name);
-            editor.putString("sso_server", account.url);
-            editor.putString("sso_type", account.type);
-            editor.putString("sso_token", account.token);
-            editor.putString("sso_userid", account.userId);
-            editor.apply();
-            setSSOPreferencesState();
-            notifyService();
-        }
-
-        private void disableSSO(){
-            SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(getContext());
-            SharedPreferences.Editor editor = sharedPreferences.edit();
-            editor.putBoolean("sso_enabled", false);
-            editor.apply();
-            setSSOPreferencesState();
-            notifyService();
-        }
-
-        private void openAccountChooser() {
-            try {
-                AccountImporter.pickNewAccount(this);
-            } catch (NextcloudFilesAppNotInstalledException | AndroidGetAccountsPermissionNotGranted e) {
-                UiExceptionManager.showDialogForException(getContext(), e);
-            }
-        }
-
-
-        private void setSSOPreferencesState(){
-            Preference login_sso = findPreference("login_sso");
-            if(login_sso == null){
-                Log.wtf(TAG, "login_sso preference is null!");
-                throw new NullPointerException();
-            }
-            if(getBoolPreference("sso_enabled",false)){
-                findPreference("server").setEnabled(false);
-                findPreference("password").setEnabled(false);
-                findPreference("login").setEnabled(false);
-                findPreference("insecure_connection").setEnabled(false);
-                login_sso.setSummary("Stop using Nextcloud app for authentication");
-                login_sso.setTitle("Log out from Nexcloud");
-                login_sso.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-                    public boolean onPreferenceClick(Preference preference) {
-                        Log.d(TAG, "Disabling SSO");
-                        disableSSO();
-                        return true;
-                    }
-                });
-            } else {
-                findPreference("server").setEnabled(true);
-                findPreference("password").setEnabled(true);
-                findPreference("login").setEnabled(true);
-                findPreference("insecure_connection").setEnabled(true);
-                login_sso.setSummary("Use on-device Nextcloud account");
-                login_sso.setTitle("Log in via Nextcloud app");
-                login_sso.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-                    public boolean onPreferenceClick(Preference preference) {
-                        Log.d(TAG, "Opening account chooser");
-                        openAccountChooser();
-                        return true;
-                    }
-                });
-            }
-        }
-
-        @Override
-        public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
-            setPreferencesFromResource(R.xml.root_preferences, rootKey);
-
-            Preference oss_licenses = (Preference) findPreference("credits");
-            oss_licenses.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-                public boolean onPreferenceClick(Preference preference) {
-                    Intent intent = new Intent(getActivity(), CreditsActivity.class);
-                    startActivity(intent);
-
-                    return true;
-                }
-            });
-
-            Preference donate = (Preference) findPreference("donate");
-            donate.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-                public boolean onPreferenceClick(Preference preference) {
-                    Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse("https://liberapay.com/Andrewerr/donate"));
-                    startActivity(browserIntent);
-                    Toast.makeText(getContext(),"Thank you!❤️", Toast.LENGTH_SHORT).show();
-                    return true;
-                }
-            });
-
-            /*
-            Preference login_sso = (Preference) findPreference("login_sso");
-            login_sso.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-                public boolean onPreferenceClick(Preference preference) {
-                    Log.d(TAG, "Opening account chooser");
-                    openAccountChooser();
-                    return true;
-                }
-            });
-             */
-
-            setSSOPreferencesState();
-        }
-
-        public void setStatus(String _status) {
-            EditTextPreference status = (EditTextPreference) findPreference("status");
-            if (status == null) {
-                Log.wtf(TAG, "Unexpected null result of findPreference");
-                throw new RuntimeException("Expected EditTextPreference, but got null!");
-            } else {
-                status.setSummary(_status);
-            }
-        }
-
-
-        @Override
-        public void onActivityResult(int requestCode, int resultCode, Intent data) {
-            super.onActivityResult(requestCode, resultCode, data);
-
-            try {
-                AccountImporter.onActivityResult(requestCode, resultCode, data, this, new AccountImporter.IAccountAccessGranted() {
-
-                    @Override
-                    public void accountAccessGranted(SingleSignOnAccount singleSignOnAccount) {
-                        enableSSO(singleSignOnAccount);
-                        Log.i(TAG, "Succesfully imported account");
-                    }
-
-                    NextcloudAPI.ApiConnectedListener callback = new NextcloudAPI.ApiConnectedListener() {
-                        @Override
-                        public void onConnected() {
-                            // ignore this one… see 5)
-                        }
-
-                        @Override
-                        public void onError(Exception ex) {
-                            ex.printStackTrace();
-                            Toast.makeText(getContext(), ex.getLocalizedMessage(), Toast.LENGTH_LONG).show();
-                        }
-                    };
-
-                });
-            } catch (AccountImportCancelledException e) {
-                e.printStackTrace();
-            }
-        }
-        @Override
-        public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-            super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-            Log.d(TAG, "Succesfully got Nextcloud permissions");
-            AccountImporter.onRequestPermissionsResult(requestCode, permissions, grantResults, this);
-        }
-
-        @Override
-        public void onDisplayPreferenceDialog(Preference preference) {
-
-            if (preference instanceof NumberDialogPreference) {
-                NumberDialogPreference dialogPreference = (NumberDialogPreference) preference;
-                DialogFragment dialogFragment = NumberPickerPreferenceDialogFragment
-                        .newInstance(
-                                dialogPreference.getKey(),
-                                dialogPreference.getMinValue(),
-                                dialogPreference.getMaxValue(),
-                                dialogPreference.getStepValue(),
-                                dialogPreference.getUnitText()
-                        );
-                dialogFragment.setTargetFragment(this, 0);
-                dialogFragment.show(getParentFragmentManager(), TAG + ".NumberPicker");
-            } else {
-                super.onDisplayPreferenceDialog(preference);
-            }
-        }
     }
