+++ app/src/main/java/com/polar/nextcloudservices/ui/donate/DonateFragment.java
@@ -1,0 +1,34 @@
+package com.polar.nextcloudservices.ui.donate;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+
+import com.polar.nextcloudservices.databinding.FragmentHomeBinding;
+
+public class DonateFragment extends Fragment {
+
+    private DonateViewModel homeViewModel;
+    private FragmentHomeBinding binding;
+
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        homeViewModel =
+                new ViewModelProvider(this).get(DonateViewModel.class);
+
+        binding = FragmentHomeBinding.inflate(inflater, container, false);
+        View root = binding.getRoot();
+        return root;
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        binding = null;
+    }
+}
