+++ app/src/main/java/com/polar/nextcloudservices/ui/home/HomeViewModel.java
@@ -1,0 +1,16 @@
+package com.polar.nextcloudservices.ui.home;
+
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
+
+public class HomeViewModel extends ViewModel {
+
+    private MutableLiveData<String> mText;
+
+    public HomeViewModel() {}
+
+    public LiveData<String> getText() {
+        return mText;
+    }
+}
