+++ app/src/main/java/com/polar/nextcloudservices/ui/settings/SettingsViewModel.java
@@ -1,0 +1,11 @@
+package com.polar.nextcloudservices.ui.settings;
+
+import androidx.lifecycle.ViewModel;
+
+public class SettingsViewModel extends ViewModel {
+
+
+    public SettingsViewModel() {
+    }
+
+}
