+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:showIn="navigation_view">
+
+    <group android:checkableBehavior="single">
+        <item
+            android:id="@+id/nav_home"
+            android:icon="@drawable/ic_menu_camera"
+            android:title="@string/menu_home" />
+        <item
+            android:id="@+id/nav_credits"
+            android:icon="@drawable/ic_menu_gallery"
+            android:title="@string/menu_credits" />
+        <item
+            android:id="@+id/nav_settings"
+            android:icon="@drawable/ic_menu_slideshow"
+            android:title="@string/menu_settings" />
+        <item
+            android:id="@+id/nav_donations"
+            android:icon="@drawable/ic_menu_slideshow"
+            android:title="@string/menu_donations" />
+    </group>
+</menu>
